[move-addr] Move Field Sensitive Pruned Liveness into its own header/cpp impl.
